Updated Camera Retrieval Method
Added a local id parameter to the camera list pulling function.  This allows the retrieval of the one camera's data, speeding up the retrieval time considerably.  Also added a button to download all HP cameras locally, replacing the single download of every camera's data from the database.  Finally, made the necessary changes for the calls to this function to work properly.