Improved 3D Model Compatability
Added a warning box to include subdirectories (will cause thumbnails to be processed as hp images) stating that the directories will not be searched for models.  Setup copyrename function to handle 3D models and their thumbnails.  This is done by copying models to <Output>\model\name-scheme\name-scheme.obj.  The preview images are placed in the same folder as the model is.  Added new parameters and new "Mandatory 3D Models" tab.  New "HP-Thumbnails" parameter is auto-populated on data processing and blocked off from the user within the HP Tool.