Updated HP-Thumbnails Generation
Changed generation of HP-Thumbnails to use the names of renamed images through the use of a dictionary.  Also added an exception to the device local ID being blank so that it will be allowed when 3D models are being processed.  Passed TypeError on validation of device local ID as it will only be hit when there is no local id, which is only allowed for 3D models anyways.