Change DirMaskSelect Plugin to Fix Bad Selection (#809)
* Change DirMaskSelect Plugin to Fix Bad Selection

Fix DirMaskSelect Plugin bug causing the plugin to always return an image from the journal directory rather than the provided directory.

Cause:
The second argument in the path join was returning an absolute path with no directory, joining the user provided directory with this would cause the absolute path to be used, and thus the image was only searched for in the journal directory, and never the user provided directory to find the image.

Fix:
Use os.path.basename on the second argument to get just the filename, with no extension and join that with the user provided directory.  Also checked if the file ended with _## as this is generally an indicator of the JT performing an operation and the filename is more likely to be found without the _##, but will still be found without it in the majority of cases

* More DirMaskSelect Fixes

Change DirMaskSelect to use copy2 from shutil instead of returning an dictionary of the new file path to avoid any name collisions.  Also updated batch_project so that the override target value is actually used.

* Revert to Previous Selection Method

Replace shutil's copy2 with the previous use of override_target and update target for override_target