Update HP Tool Release Form Capability and Fix Bugs (#833)
Update HP Tool release forms to allow copy, paste, and column fill functionality.
This did not previously work as the column has all cells in the disabled cells group, which has a blocked copy, paste, and fill functionalities.  Solved by allowing data to only be pasted into HP-ReleaseForms, to be from other HP-ReleaseForms as well as adding the condition that the column is not HP-ReleaseForms to the check if column fill is allowed.

Fix undo/redo bug with invalid redo and undo types.
Undo was passed as an instance instead of a list when no changes have occured yet.

Fix bug allowing for deletion of cells that are disabled.
Using the built in pandastable deleteCells function allowed for disabled cells to be cleared with no undo.  deleteCells function was overriden with a function that verifies that the values are allowed to be deleted, and implemented an undo to cell data deletion.