Add check output type nitf (#842)
* HP Tool Fixes (#832)

HPSpreadsheet - Disable "HP-Keywords" column on all spreadsheets that are not of models.  Except ReadError so a new archive can be created when a corrupted one is found.

test_hp_tool - Eliminated possibility of legitimate .maskgen2 usage and forced Mock to fix concatenation error.

* Update HP Tool Release Form Capability and Fix Bugs (#833)

Update HP Tool release forms to allow copy, paste, and column fill functionality.
This did not previously work as the column has all cells in the disabled cells group, which has a blocked copy, paste, and fill functionalities.  Solved by allowing data to only be pasted into HP-ReleaseForms, to be from other HP-ReleaseForms as well as adding the condition that the column is not HP-ReleaseForms to the check if column fill is allowed.

Fix undo/redo bug with invalid redo and undo types.
Undo was passed as an instance instead of a list when no changes have occured yet.

Fix bug allowing for deletion of cells that are disabled.
Using the built in pandastable deleteCells function allowed for disabled cells to be cleared with no undo.  deleteCells function was overriden with a function that verifies that the values are allowed to be deleted, and implemented an undo to cell data deletion.

* Small fixes (#831)

* Small Fixes

handling for if username or organization are None in hp_data.process

fix check length smaller and larger bug that always compared 0 with 0

fix references to cv2 Color Conversion attributes in mediated compare

remove End Time parameter from Paste Frames since jt calculates end time automagically.

* version Bump

* Fix Batch Export

* Batch Improvements (#836)

* Improve JSONSelector Error Message

Improve JSONSelector error message to tell the user that the indexed item does not exist in the JSON rather than throwing an error because the dictionary is not defined.

* Allow Predefined Description on Plugin Operations

Allow the specification to define a description for operation links created by a plugin rather than defaulting to the plugin's description

Format filename with global state

* Update Exif Changing Plugins

exif.py - Add get_version command to retrieve exiftool full version string

ExifDateTime - Change command from ffmpeg to exiftool.  This allows for dates to be added when they don't exist rather than just updating them.  Allowed - and / date seperators.  Fix faulty if statements.

ExifGPSChange - Fix faulty logic with ok variable being initialized to True, leading it to always be true.  Changed to get_version command from maskgen.exif instead of hardcoded (and outdated) exiftool version.

* Split last item off tool path (#837)

prevent user paths being recorded, just record tool/script.

* fix for taskDesignation (#838)

* Fix HP Tool Test (#835)

Fix HP Tool Test to allow for the changes made to hp_data by Luke

* Adding ImageStabilizer Keywords for HP Tool (#834)

Adding ImageStabilizer Keywords for HP Tool

* fix typo (#840)

* Add checkOutputTypeNITF

add specific checkOutput function to allow for .ntf extension