Added logging to ExifDateTime Plugin + Fixed Logic
Added log statements to ExifDateTime plugin to warn the user which date and time are being used as a default and what the date/time is being changed to. This is so there is no question of whether the current date/time or the date/time from the exif data was used for a default, and confirmation that the date and time being set are valid.

Fixed logic for date randomization as information that was not supposed to be randomized, was being randomized because None is not 'None'


Former-commit-id: dc4c9964bf6a371649a79a8301ddac3b97a85516